Fix,refact/76/chat room

+++ src/chat-room/chat-room.service.ts
@@ -10,6 +10,7 @@ import { ChatRoomUserRepository } from './repository/chatRoomUser.repository';
 import { ChatRoomRole } from './enum/chat-room-role.enum';
 import { ChatRoomUserEntity } from './entities/chatRoomUser.entity';
 import { Socket } from 'socket.io';
+import { ErrorStatus } from './enum/error-status.enum';
 
 @Injectable()
 export class ChatRoomService {
@@ -33,18 +34,25 @@ export class ChatRoomService {
     chatRoom: ChatRoomEntity,
     user: UserEntity,
     role: ChatRoomRole,
-  ): Promise<void> {
+  ): Promise<ChatRoomUserEntity> {
     const chatRoomUser = await this.chatRoomUserRepository.getChatRoomUser(
       chatRoom,
       user,
     );
     if (chatRoomUser) {
       if (chatRoomUser.isBanned) {
-        throw new WsException('해당 채팅방에서 차단당한 유저입니다.');
+        throw new WsException({
+          status: ErrorStatus.ERROR,
+          message: '해당 채팅방에서 차단당한 유저입니다.',
+        });
       }
-      return;
+      return chatRoomUser;
     }
-    await this.chatRoomUserRepository.createChatRoomUser(chatRoom, user, role);
+    return await this.chatRoomUserRepository.createChatRoomUser(
+      chatRoom,
+      user,
+      role,
+    );
   }
 
   async getChatRoomUser(chatRoom: ChatRoomEntity, user: UserEntity) {
@@ -55,6 +63,12 @@ export class ChatRoomService {
     chatRoom: ChatRoomEntity,
   ): Promise<ChatRoomUserEntity[]> {
     return await this.chatRoomUserRepository.getChatRoomUsers(chatRoom);
+  }
+
+  async getChatRoomUsersExceptBanned(chatRoom: ChatRoomEntity) {
+    return await this.chatRoomUserRepository.getChatRoomUsersExceptBanned(
+      chatRoom,
+    );
   }
 
   async toggleBanUser(chatRoomUser: ChatRoomUserEntity): Promise<void> {
@@ -70,6 +84,10 @@ export class ChatRoomService {
     chatRoomRole: ChatRoomRole,
   ): Promise<void> {
     chatRoomUser.role = chatRoomRole;
+    if (chatRoomRole <= ChatRoomRole.ADMIN) {
+      chatRoomUser.isMuted = false;
+      chatRoomUser.mutedUntil = null;
+    }
     await this.chatRoomUserRepository.saveChatRoomUser(chatRoomUser);
   }
 
@@ -77,10 +95,11 @@ export class ChatRoomService {
     chatRoomUser: ChatRoomUserEntity,
     isMuted: boolean,
   ): Promise<void> {
+    const now = new Date();
     chatRoomUser.isMuted = isMuted;
     chatRoomUser.mutedUntil = null;
     if (isMuted) {
-      chatRoomUser.mutedUntil = new Date();
+      chatRoomUser.mutedUntil = new Date(now.getTime() + 1000 * 60 * 5);
     }
     await this.chatRoomUserRepository.saveChatRoomUser(chatRoomUser);
   }
@@ -108,13 +127,18 @@ export class ChatRoomService {
     payload: string,
   ): Promise<any> {
     if (!payload || payload.length >= 1000) {
-      throw new WsException('메세지가 비어있거나 너무 큽니다.');
+      throw new WsException({
+        status: ErrorStatus.WARNING,
+        message: '메세지가 비어있거나 너무 큽니다.',
+      });
     }
     if (chatRoomUser.isMuted) {
       const now = new Date();
-      // NOTE: 5 minutes
-      if (now.getTime() <= chatRoomUser?.mutedUntil.getTime() + 1000 * 60 * 5) {
-        throw new WsException('관리자에 의해 mute 되었습니다.');
+      if (now.getTime() <= chatRoomUser?.mutedUntil.getTime()) {
+        throw new WsException({
+          status: ErrorStatus.WARNING,
+          message: '관리자에 의해 mute 되었습니다.',
+        });
       }
       await this.setMuteUser(chatRoomUser, false);
     }
@@ -144,8 +168,11 @@ export class ChatRoomService {
     if (!chatRoomUser) {
       return;
     }
+    if (chatRoomUser.isBanned === true) {
+      return;
+    }
     await this.chatRoomUserRepository.deleteChatRoomUser(chatRoomUser);
-    const chatRoomUsers = await this.getChatRoomUsers(chatRoom);
+    const chatRoomUsers = await this.getChatRoomUsersExceptBanned(chatRoom);
 
     // NOTE: 만약 나가는 유저가 그방의 마지막 유저면 방을 삭제해야함
     if (chatRoomUsers.length === 0) {
